refactor(cloud): extract viewer title and presence chrome
