+++ apps/notebook-cloud/test/viewer-shared-cell-surface.test.ts
@@ -264,22 +264,26 @@ test("cloud identity chrome renders through the shared actor projection surface"
 test("cloud presence chrome renders as an isolated host avatar stack", () => {
   const sourcePath = new URL("../viewer/index.tsx", import.meta.url);
   const sourceText = readFileSync(sourcePath, "utf8");
+  const presenceSourcePath = new URL("../viewer/cloud-presence-status.tsx", import.meta.url);
+  const presenceSourceText = readFileSync(presenceSourcePath, "utf8");
   const hostedSmokePath = new URL("../scripts/hosted-render-smoke.mjs", import.meta.url);
   const hostedSmokeText = readFileSync(hostedSmokePath, "utf8");
   const collabSmokePath = new URL("../scripts/hosted-collab-smoke.mjs", import.meta.url);
   const collabSmokeText = readFileSync(collabSmokePath, "utf8");
 
-  assert.match(sourceText, /CloudViewerPresenceStore/);
+  assert.match(presenceSourceText, /CloudViewerPresenceStore/);
   assert.match(
-    sourceText,
+    presenceSourceText,
     /useSyncExternalStore\(store\.subscribe, store\.getSnapshot, store\.getSnapshot\)/,
   );
-  assert.match(sourceText, /<AvatarGroup className="cloud-presence-avatar-group"/);
-  assert.match(sourceText, /data-slot="cloud-presence-stack"/);
+  assert.match(presenceSourceText, /<AvatarGroup className="cloud-presence-avatar-group"/);
+  assert.match(presenceSourceText, /data-slot="cloud-presence-stack"/);
   assert.match(hostedSmokeText, /\[data-slot='cloud-presence-stack'\]/);
   assert.match(collabSmokeText, /\[data-slot='cloud-presence-stack'\]/);
   assert.doesNotMatch(sourceText, /NotebookPresenceStatus/);
+  assert.doesNotMatch(presenceSourceText, /NotebookPresenceStatus/);
   assert.doesNotMatch(sourceText, /compactCloudPresenceLabel/);
+  assert.doesNotMatch(presenceSourceText, /compactCloudPresenceLabel/);
   assert.doesNotMatch(hostedSmokeText, /notebook-presence-status/);
   assert.doesNotMatch(collabSmokeText, /notebook-presence-status/);
 });
