+++ apps/notebook-cloud/viewer/cloud-presence-status.tsx
@@ -1,0 +1,115 @@
+import { useSyncExternalStore } from "react";
+import { UserRound } from "lucide-react";
+import {
+  Avatar,
+  AvatarBadge,
+  AvatarFallback,
+  AvatarGroup,
+  AvatarGroupCount,
+} from "@/components/ui/avatar";
+import {
+  cloudViewerPresenceDisplay,
+  type CloudViewerPresencePeer,
+  type CloudViewerPresenceStore,
+} from "./presence";
+
+export function CloudPresenceStatus({
+  connectionError,
+  store,
+}: {
+  connectionError: string | null;
+  store: CloudViewerPresenceStore;
+}) {
+  const presence = useSyncExternalStore(store.subscribe, store.getSnapshot, store.getSnapshot);
+  const presenceDisplay = cloudViewerPresenceDisplay(presence);
+  const connected = presenceDisplay.connected && !connectionError;
+  const title = connectionError
+    ? `Room unavailable: ${cloudConnectionStatusErrorTitle(connectionError)}`
+    : presenceDisplay.title;
+  const state = connected ? "live" : presence.connection === "connecting" ? "joining" : "waiting";
+
+  return (
+    <span
+      className="cloud-presence-stack"
+      data-slot="cloud-presence-stack"
+      data-state={state}
+      title={title}
+      aria-label={title}
+    >
+      <AvatarGroup className="cloud-presence-avatar-group" aria-hidden="true">
+        {presenceDisplay.peers.map((peer) => (
+          <CloudPresenceAvatar key={peer.id} peer={peer} connected={connected} />
+        ))}
+        {presenceDisplay.hiddenCount > 0 ? (
+          <AvatarGroupCount className="cloud-presence-avatar-count" data-size="sm">
+            +{presenceDisplay.hiddenCount}
+          </AvatarGroupCount>
+        ) : null}
+      </AvatarGroup>
+      <span className="sr-only">{presenceDisplay.label}</span>
+    </span>
+  );
+}
+
+function cloudConnectionStatusErrorTitle(error: string): string {
+  if (/\bfailed to connect\s+wss?:\/\//i.test(error)) {
+    return "unable to join the live room";
+  }
+  return sanitizeCloudConnectionError(error);
+}
+
+function sanitizeCloudConnectionError(error: string): string {
+  return error.replace(/\bwss?:\/\/[^\s]+/gi, (rawUrl) => {
+    try {
+      const url = new URL(rawUrl);
+      return `${url.protocol}//${url.host}${url.pathname}`;
+    } catch {
+      return rawUrl.replace(/[?#].*$/, "");
+    }
+  });
+}
+
+function CloudPresenceAvatar({
+  connected,
+  peer,
+}: {
+  connected: boolean;
+  peer: CloudViewerPresencePeer;
+}) {
+  const status = connected ? peer.status : "offline";
+  return (
+    <Avatar
+      size="sm"
+      className="cloud-presence-avatar"
+      data-kind={peer.kind}
+      data-status={status}
+      title={peer.label}
+    >
+      <AvatarFallback>
+        {peer.kind === "anonymous" ? (
+          <>
+            <UserRound aria-hidden="true" />
+            {peer.count && peer.count > 1 ? (
+              <span className="cloud-presence-anonymous-count">{peer.count}</span>
+            ) : null}
+          </>
+        ) : (
+          cloudPresenceInitials(peer.label)
+        )}
+      </AvatarFallback>
+      <AvatarBadge data-status={status} />
+    </Avatar>
+  );
+}
+
+function cloudPresenceInitials(label: string): string {
+  const words = label
+    .split(/[\s@._-]+/g)
+    .map((word) => word.trim())
+    .filter(Boolean);
+  const initials = words
+    .slice(0, 2)
+    .map((word) => word[0]?.toUpperCase() ?? "")
+    .join("");
+  return initials || "?";
+}
